@@ -4,40 +4,46 @@
     "type": "project",
     "require": {
         "php": ">=5.6.0",
-        "symfony/symfony": "~3.3.0",
-        "defuse/php-encryption": "~2.0.1",
-        "doctrine/dbal": "~2.5.3",
-        "doctrine/cache": "~1.6|~1.5",
-        "doctrine/common": "~2.5.3",
-        "doctrine/orm": "~2.5.3",
         "ext-curl": "*",
         "ext-intl": "*",
         "ext-zip": "*",
-        "geoip2/geoip2": "~2.4.2",
+        "curl/curl": "^1.2.1",
+        "symfony/polyfill-apcu": "^1.0",
+        "symfony/symfony": "^3.4.0",
+        "symfony/swiftmailer-bundle": "^2.6.4",
+        "symfony/monolog-bundle": "^3.1.0",
+        "doctrine/dbal": "^2.5.3",
+        "doctrine/cache": "^1.6",
+        "doctrine/common": "~2.5.3",
+        "doctrine/orm": "~2.5.3",
         "doctrine/doctrine-bundle": "^1.6",
+        "twig/twig": "^1.0||^2.0",
+        "smarty/smarty": "^3.1.31",
+        "swiftmailer/swiftmailer": "~5.4.3",
+        "sensio/distribution-bundle": "^5.0.19",
+        "sensio/framework-extra-bundle": "^5.0.0",
+        "defuse/php-encryption": "~2.0.1",
+        "geoip2/geoip2": "~2.4.2",
         "mrclay/minify": "~2.3.0",
         "matthiasmullie/minify": "~1.3.0",
         "jakeasmith/http_build_url": "^1",
         "mobiledetect/mobiledetectlib": "~2.8.22",
-        "symfony/swiftmailer-bundle": "^2.3.10",
         "pear/archive_tar": "~1.4.2",
-        "swiftmailer/swiftmailer": "~5.4.3",
-        "symfony/monolog-bundle": "^3.1.0",
-        "sensio/distribution-bundle": "^5.0.19",
-        "sensio/framework-extra-bundle": "^3.0.2",
-        "prestashop/translationtools-bundle": "~2.0",
-        "tecnickcom/tcpdf": "6.2.12",
-        "composer/installers": "1.0.21",
+        "tecnickcom/tcpdf": "^6.2.12",
+        "composer/installers": "^1.0.21",
         "icanboogie/cldr": "1.3.9",
         "incenteev/composer-parameter-handler": "~2.0",
-        "curl/curl": "1.2.1",
-        "ircmaxell/password-compat": "1.0.4",
-        "ircmaxell/random-lib": "*",
-        "shudrum/array-finder": "1.1.0",
+        "ircmaxell/password-compat": "^1.0.4",
+        "ircmaxell/random-lib": "^1.2.0",
+        "shudrum/array-finder": "^1.1.0",
         "phpoffice/phpexcel": "~1.8",
         "guzzlehttp/guzzle": "~5.0",
         "csa/guzzle-bundle": "~1.3",
         "paragonie/random_compat": "^1.4|^2.0",
+        "beberlei/DoctrineExtensions": "^1.0",
+        "composer/ca-bundle": "^1.0",
+        "nikic/php-parser": "^2.1",
+        "prestashop/translationtools-bundle": "~2.0",
         "prestashop/blockreassurance": "^3",
         "prestashop/contactform": "^3",
         "prestashop/dashactivity": "^2",
