@@ -30,28 +30,99 @@ class CurrencyCore extends ObjectModel
 {
     public $id;
 
-    /** @var string Name */
+    /**
+     * Name of the currency
+     *
+     * @var string
+     */
     public $name;
 
-    /** @var string Iso code */
+    /**
+     * Alphabetic ISO 4217 code of this currency
+     *
+     * @var string
+     */
     public $iso_code;
 
-    /** @var string numeric Iso code */
+    /**
+     * Numeric ISO 4217 code of this currency
+     * Will be deprecated soon
+     *
+     * @var string
+     */
     public $iso_code_num;
 
-    /** @var string exchange rate from euros */
+    /**
+     * Numeric ISO 4217 code of this currency
+     *
+     * @var string
+     */
+    public $numeric_iso_code;
+
+    /**
+     * Exchange rate from default currency
+     *
+     * @var float
+     */
     public $conversion_rate;
 
-    /** @var bool True if currency has been deleted (staying in database as deleted) */
+    /**
+     * Is this currency deleted ?
+     * If currency is deleted, it stays in database. This is just a state (soft delete).
+     *
+     * @var bool
+     */
     public $deleted = 0;
 
-    /** @var int bool active */
+    /**
+     * Is this currency active ?
+     *
+     * @var int|bool active
+     */
     public $active;
 
+    /**
+     * Currency's symbol
+     * Will be deprecated soon
+     *
+     * @var string
+     */
     public $sign;
+
+    /**
+     * Currency's symbol
+     *
+     * @var string
+     */
+    public $symbol;
+
+    /**
+     * CLDR price formatting pattern
+     * e.g.: In french (fr-FR), price formatting pattern is : #,##0.00 ¤
+     *
+     * @var string
+     */
     public $format;
+
+    /**
+     * @var int
+     */
     public $blank;
+
+    /**
+     * Number of decimal digits to use when displaying a price in this currency
+     * Will be deprecated soon
+     *
+     * @var int
+     */
     public $decimals;
+
+    /**
+     * Number of decimal digits to use when displaying a price in this currency
+     *
+     * @var int
+     */
+    public $precision;
 
     /**
      * @see ObjectModel::$definition
@@ -59,13 +130,19 @@ class CurrencyCore extends ObjectModel
     public static $definition = array(
         'table' => 'currency',
         'primary' => 'id_currency',
-        'multilang_shop' => true,
+        'multilang' => true,
+        // 'multilang_shop' => true,
         'fields' => array(
-            'name' => array('type' => self::TYPE_STRING, 'validate' => 'isGenericName', 'required' => true, 'size' => 64),
-            'iso_code' => array('type' => self::TYPE_STRING, 'validate' => 'isLanguageIsoCode', 'required' => true, 'size' => 3),
-            'conversion_rate' => array('type' => self::TYPE_FLOAT, 'validate' => 'isUnsignedFloat', 'required' => true, 'shop' => true),
-            'deleted' => array('type' => self::TYPE_BOOL, 'validate' => 'isBool'),
-            'active' => array('type' => self::TYPE_BOOL, 'validate' => 'isBool'),
+            'iso_code'         => array('type' => self::TYPE_STRING, 'validate' => 'isLanguageIsoCode', 'required' => true, 'size' => 3),
+            'numeric_iso_code' => array('type' => self::TYPE_STRING, 'validate' => 'isNumericIsoCode', 'size' => 3),
+            'precision'        => array('type' => self::TYPE_INT   , 'validate' => 'isInt'),
+            'conversion_rate'  => array('type' => self::TYPE_FLOAT , 'validate' => 'isUnsignedFloat'  , 'required' => true, 'shop' => true),
+            'deleted'          => array('type' => self::TYPE_BOOL  , 'validate' => 'isBool'),
+            'active'           => array('type' => self::TYPE_BOOL  , 'validate' => 'isBool'),
+
+            /* Lang fields */
+            'name'   => array('type' => self::TYPE_STRING, 'lang' => true, 'validate' => 'isGenericName', 'size' => 255),
+            'symbol' => array('type' => self::TYPE_STRING, 'lang' => true                               , 'size' => 255),
         ),
     );
 
@@ -104,12 +181,14 @@ class CurrencyCore extends ObjectModel
         if ($this->iso_code) {
             $cldrCurrency = $this->cldr->getCurrency($this->iso_code);
 
-            $this->sign = $cldrCurrency['symbol'];
-            $this->iso_code_num = $cldrCurrency['iso_code'];
-            $this->name = $cldrCurrency['name'];
-            $this->format = $this->cldr->getCurrencyFormatPattern();
-            $this->blank = 1;
-            $this->decimals = 1;
+            $this->sign             = $cldrCurrency['symbol'];
+            $this->symbol           = $cldrCurrency['symbol'];
+            $this->iso_code_num     = $cldrCurrency['iso_code'];
+            $this->numeric_iso_code = $cldrCurrency['iso_code'];
+            $this->name             = $cldrCurrency['name'];
+            $this->format           = $this->cldr->getCurrencyFormatPattern();
+            $this->blank            = 1;
+            $this->decimals         = 1;
         }
 
         if (!$this->conversion_rate) {
