@@ -30,6 +30,11 @@ namespace PrestaShop\PrestaShop\Core\Localization\CLDR;
 use PrestaShop\PrestaShop\Core\Localization\Exception\LocalizationException;
 use SimpleXMLElement;
 
+/**
+ * CLDR files reader class
+ *
+ * This class provides CLDR LocaleData objects built with data coming from official CLDR xml data files.
+ */
 class Reader implements ReaderInterface
 {
     const CLDR_ROOT         = 'localization/CLDR/';
@@ -54,6 +59,13 @@ class Reader implements ReaderInterface
      * @var SimplexmlElement
      */
     protected $supplementalXml;
+
+    /**
+     * Additional data about numbering systems
+     * Mainly used for digits (they depend on numbering system)
+     *
+     * @var SimpleXMLElement
+     */
     protected $numberingSystemsXml;
 
     /**
@@ -93,7 +105,7 @@ class Reader implements ReaderInterface
      * Validate a locale code
      *
      * If the passed code doesn't respect the CLDR files naming style, an exception will be raised
-     * eg : "fr_FR" and "en_001" are valid
+     * e.g.: "fr_FR" and "en_001" are valid
      *
      * @param $localeCode
      *  Locale code to be validated
@@ -202,28 +214,8 @@ class Reader implements ReaderInterface
             return $parent;
         }
 
-        // The "top level" case. When only language code is left in $localeCode : 'en', 'fr'... then parent is "root".
+        // The "top level" case. When only language code is left in $localeCode: 'en', 'fr'... then parent is "root".
         return self::CLDR_ROOT_LOCALE;
-    }
-
-    /**
-     * Extracts the relevant parts of an IETF locale tag
-     *
-     * @param string $localeTag The locale tag (e.g.: fr-FR, en-US...)
-     *
-     * @return array The indexed locale parts (e.g.: ['langage' => 'en', 'region' => 'US'])
-     */
-    protected function getLocaleParts($localeTag)
-    {
-        $expl  = explode('-', $localeTag);
-        $parts = [
-            'language' => $expl[0],
-        ];
-        if (!empty($expl[1])) {
-            $parts['region'] = $expl[1];
-        }
-
-        return $parts;
     }
 
     /**
@@ -300,12 +292,16 @@ class Reader implements ReaderInterface
     protected function mapLocaleData(SimplexmlElement $xmlLocaleData, $supplementalData)
     {
         $localeData = new LocaleData();
+
+        // Geo
         if (isset($xmlLocaleData->identity->language)) {
             $localeData->localeCode = (string)$xmlLocaleData->identity->language['type'];
         }
         if (isset($xmlLocaleData->identity->territory)) {
             $localeData->localeCode .= '-' . $xmlLocaleData->identity->territory['type'];
         }
+
+        // Numbers
         $numbersData = $xmlLocaleData->numbers;
         // Default numbering system.
         if (isset($numbersData->defaultNumberingSystem)) {
@@ -398,7 +394,7 @@ class Reader implements ReaderInterface
             foreach ($numbersData->decimalFormats as $format) {
                 /** @var SimplexmlElement $format */
                 $numberSystem = (string)$format['numberSystem'];
-                // If alias is set, we just copy data from another numbering system :
+                // If alias is set, we just copy data from another numbering system:
                 $alias = $format->alias;
                 if ($alias
                     && preg_match(
@@ -428,7 +424,7 @@ class Reader implements ReaderInterface
             foreach ($numbersData->percentFormats as $format) {
                 /** @var SimplexmlElement $format */
                 $numberSystem = (string)$format['numberSystem'];
-                // If alias is set, we just copy data from another numbering system :
+                // If alias is set, we just copy data from another numbering system:
                 $alias = $format->alias;
                 if ($alias
                     && preg_match(
@@ -459,7 +455,7 @@ class Reader implements ReaderInterface
             foreach ($numbersData->currencyFormats as $format) {
                 /** @var SimplexmlElement $format */
                 $numberSystem = (string)$format['numberSystem'];
-                // If alias is set, we just copy data from another numbering system :
+                // If alias is set, we just copy data from another numbering system:
                 $alias = $format->alias;
                 if ($alias
                     && preg_match(
@@ -475,37 +471,59 @@ class Reader implements ReaderInterface
             }
         }
 
+        // Currencies
+        $currencyData = $numbersData->currencies;
+        if (isset($currencyData->currency)) {
+            foreach ($currencyData->currency as $currencyNode) {
+                $currencyCode = (string)$currencyNode['type'];
+
+                $currencyData          = new CurrencyData();
+                $currencyData->isoCode = $currencyCode;
+
+                // Symbols
+                foreach ($currencyNode->symbol as $symbolNode) {
+                    $type = (string)$symbolNode['alt'];
+                    if (empty($type)) {
+                        $type = 'default';
+                    }
+                    $currencyData->symbols[$type] = (string)$symbolNode;
+                }
+
+                // Names
+                foreach ($currencyNode->displayName as $nameNode) {
+                    $countContext = 'default';
+                    if (!empty($nameNode['count'])) {
+                        $countContext = (string)$nameNode['count'];
+                    }
+                    $currencyData->displayNames[$countContext] = (string)$nameNode;
+                }
+
+                // Supplemental (fraction digits and numeric iso code)
+                $codesMapping = $this->supplementalXml->supplementalData->xpath(
+                    '//codeMappings/currencyCodes[@type="' . $currencyCode . '"]'
+                );
+
+                if (!empty($codesMapping)) {
+                    /** @var SimplexmlElement $codesMapping */
+                    $codesMapping                 = $codesMapping[0];
+                    $currencyData->numericIsoCode = (string)$codesMapping->attributes()->numeric;
+                }
+
+                $fractionsData = $this->supplementalXml->supplementalData->xpath(
+                    '//currencyData/fractions/info[@iso4217="' . $currencyCode . '"]'
+                );
+
+                if (!empty($fractionsData)) {
+                    /** @var SimplexmlElement $fractionsData */
+                    $fractionsData               = $fractionsData[0];
+                    $currencyData->decimalDigits = (int)(string)$fractionsData->attributes()->digits;
+                }
+
+                $localeData->currencies[$currencyCode] = $currencyData;
+            }
+        }
+
         return $localeData;
-    }
-
-
-
-
-    /**
-     * Extract parent locale code
-     *
-     * @param SimplexmlElement $parentLocaleXmlData
-     * @param string $localeTag
-     *
-     * @return mixed|string
-     */
-    protected function extractParentLocale(SimplexmlElement $parentLocaleXmlData, $localeTag)
-    {
-        if (self::CLDR_ROOT_LOCALE === $localeTag) {
-            return null;
-        }
-        $parts = $this->getLocaleParts($localeTag);
-        if (empty($parts['region'])) {
-            return self::CLDR_ROOT_LOCALE;
-        }
-        $code    = $parts['language'] . '_' . $parts['region'];
-        $results = $parentLocaleXmlData->xpath("//parentLocale[contains(@locales, '$code')]");
-        if (empty($results)) {
-            return $parts['language'];
-        }
-        $node = $results[0];
-
-        return (string)$node['parent'];
     }
 
     /**
