@@ -47,7 +47,7 @@ class Currency implements CurrencyInterface
      *
      * Price in currency A * currency A's conversion rate = price in default currency
      *
-     * Example :
+     * Example:
      * Given the Euro as default shop's currency,
      * If 1 dollar = 1.31 euros,
      * Then conversion rate for Dollar will be 1.31
@@ -119,7 +119,7 @@ class Currency implements CurrencyInterface
      * @param int $precision
      *  Number of decimal digits to use with this currency
      *
-     * @param string [] $names
+     * @param string[] $names
      *  the currency's name, by locale code
      *
      */
