@@ -106,7 +106,7 @@ class Formatter
         try {
             $decimalNumber = $this->prepareNumber($number);
         } catch (SPLInvalidArgumentException $e) {
-            throw new LocalizationException('Invalid $number parameter : ' . $e->getMessage(), 0, $e);
+            throw new LocalizationException('Invalid $number parameter: ' . $e->getMessage(), 0, $e);
         }
 
         /*
@@ -162,9 +162,9 @@ class Formatter
      * Get $number's major and minor digits.
      *
      * Major digits are the "integer" part (before decimal separator), minor digits are the fractional part
-     * Result will be an array of exactly 2 items : [$majorDigits, $minorDigits]
+     * Result will be an array of exactly 2 items: [$majorDigits, $minorDigits]
      *
-     * Usage example :
+     * Usage example:
      *  list($majorDigits, $minorDigits) = $this->getMajorMinorDigits($decimalNumber);
      *
      * @param DecimalNumber $number
@@ -184,7 +184,7 @@ class Formatter
     /**
      * Splits major digits into groups
      *
-     * eg. : Given the major digits "1234567", and major group size
+     * e.g.: Given the major digits "1234567", and major group size
      *  configured to 3 digits, the result would be "1 234 567"
      *
      * @param $majorDigits
@@ -332,9 +332,9 @@ class Formatter
      *
      * Missing placeholders can be the percent sign, currency symbol, etc.
      *
-     * e.g. with a currency CLDR pattern :
-     *  - Passed number (partially formatted) : 1,234.567
-     *  - Returned number : 1,234.567 ¤
+     * e.g. with a currency CLDR pattern:
+     *  - Passed number (partially formatted): 1,234.567
+     *  - Returned number: 1,234.567 ¤
      *  ("¤" symbol is the currency symbol placeholder)
      *
      * @see http://cldr.unicode.org/translation/number-patterns
@@ -350,7 +350,7 @@ class Formatter
     protected function addPlaceholders($formattedNumber, $pattern)
     {
         /*
-         * Regex groups explanation :
+         * Regex groups explanation:
          * #          : literal "#" character. Once.
          * (,#+)*     : any other "#" characters group, separated by ",". Zero to infinity times.
          * 0          : literal "0" character. Once.
