@@ -37,12 +37,12 @@ use PrestaShop\PrestaShop\Core\Localization\Specification\Number as NumberSpecif
 class Price extends NumberSpecification
 {
     /**
-     * Currency display option : symbol notation
+     * Currency display option: symbol notation
      */
     const CURRENCY_DISPLAY_SYMBOL = 'symbol';
 
     /**
-     * Currency display option : ISO code notation
+     * Currency display option: ISO code notation
      */
     const CURRENCY_DISPLAY_CODE = 'code';
 
@@ -56,13 +56,13 @@ class Price extends NumberSpecification
 
     /**
      * @var string The currency symbol
-     * eg : €
+     * e.g.: €
      */
     protected $currencySymbol;
 
     /**
      * @var string The currency code
-     * eg : EUR
+     * e.g.: EUR
      */
     protected $currencyCode;
 
@@ -98,10 +98,10 @@ class Price extends NumberSpecification
      *  Type of display for currency symbol
      *
      * @param string $currencySymbol
-     *  Currency symbol of this price (eg. : €)
+     *  Currency symbol of this price (e.g.: €)
      *
      * @param $currencyCode
-     *  Currency code of this price (eg. : EUR)
+     *  Currency code of this price (e.g.: EUR)
      *
      * @throws LocalizationException
      */
@@ -146,7 +146,7 @@ class Price extends NumberSpecification
 
     /**
      * Get the currency symbol
-     * eg. : €
+     * e.g.: €
      *
      * @return string
      */
@@ -157,7 +157,7 @@ class Price extends NumberSpecification
 
     /**
      * Get the currency ISO code
-     * eg. : EUR
+     * e.g.: EUR
      *
      * @return string
      */
