@@ -2,9 +2,9 @@ const {AccessPageBO} = require('../../../selectors/BO/access_page');
 const {OrderPage} = require('../../../selectors/BO/order');
 const {CreateOrder} = require('../../../selectors/BO/order');
 const {OnBoarding} = require('../../../selectors/BO/onboarding.js');
-const orderScenarios = require('./order');
+const orderScenarios = require('../../common_scenarios/order');
 const {Menu} = require('../../../selectors/BO/menu.js');
-const common_scenarios = require('../02_product/product');
+const common_scenarios = require('../../common_scenarios/product');
 const {AddProductPage} = require('../../../selectors/BO/add_product_page');
 let promise = Promise.resolve();
 
