@@ -2,14 +2,14 @@ var CommonClient = require('./common_client');
 
 class AttributeAndFeature extends CommonClient {
 
-  addValueToAttribute(attributeSubMenu) {
+  addValueToAttribute(attributeSubMenu, values) {
     return this.client
       .waitForExistAndClick(attributeSubMenu.add_value_button)
-      .waitAndSetValue(attributeSubMenu.value_input, "10")
+      .waitAndSetValue(attributeSubMenu.value_input, values[1])
       .waitForExistAndClick(attributeSubMenu.save_and_add)
-      .waitAndSetValue(attributeSubMenu.value_input, "20")
+      .waitAndSetValue(attributeSubMenu.value_input, values[2])
       .waitForExistAndClick(attributeSubMenu.save_and_add)
-      .waitAndSetValue(attributeSubMenu.value_input, "30")
+      .waitAndSetValue(attributeSubMenu.value_input, values[3])
       .waitForExistAndClick(attributeSubMenu.save_value_button)
   }
 
