@@ -5,7 +5,7 @@ module.exports = {
     password_inputFO: '//*[@id="login-form"]/section/div[2]/div[1]/div/input',
     login_button: '//*[@id="login-form"]/footer/button',
     sign_out_button: '//*[@id="_desktop_user_info"]/div/a[1]',
-    logo_home_page: '//*[@id="_desktop_logo"]/a',
+    logo_home_page: '//*[@id="_desktop_logo"]//a',
     product_list_button: '//*[@id="content"]/section/a',
     categories_list: '//*[@id="left-column"]/div[1]/ul/li[2]/ul',
     shopping_cart_button: '//*[@id="_desktop_cart"]/div',
