@@ -28,7 +28,7 @@ namespace Tests\TestCase;
 
 use PHPUnit\Framework\TestCase;
 use Tests\Unit\ContextMocker;
-use Tests\PrestaShopBundle\Utils\Database;
+use Tests\PrestaShopBundle\Utils\DatabaseCreator as Database;
 
 class IntegrationTestCase extends TestCase
 {
